requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
build==1.2.2
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==43.0.1
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastjsonschema==2.20.0
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
Mako==1.3.5
MarkupSafe==3.0.1
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.1
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomlkit==0.13.2
trove-classifiers==2024.9.12
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
virtualenv==20.26.5
watchfiles==0.24.0
websockets==13.1